Preserve same instances of NavEntry after restoration from Parcel
If there have been two or more items of the same instance before parceling, there should exactly the same items of the same instance after recreation from Parcel. This would be prevent a group of potential unnoying issues for developers who doesn't understand ins and outs of the library. And just simplifies things.

This also resolves some unnecessary notes in the documentation and in the sample.